% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-barmkekko.R
\docType{data}
\name{GeomBarMekko}
\alias{GeomBarMekko}
\title{GeomBarMekko}
\format{An object of class \code{GeomBarMekko} (inherits from \code{GeomRect}, \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
GeomBarMekko
}
\description{
GeomBarMekko
}
\keyword{datasets}
