% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barmekko.R
\name{positions}
\alias{positions}
\title{Calculate positions from widths.}
\usage{
positions(width)
}
\arguments{
\item{width}{A numeric vector of bar widths.}
}
\value{
A numeric vector of bar positions.
}
\description{
Calculate positions from widths.
}

