% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marimekko.R
\name{marimekko}
\alias{marimekko}
\title{Create a Marimekko plot.}
\usage{
marimekko(df, x, y, width)
}
\arguments{
\item{df}{A data frame.}

\item{x}{A categorical variable defining the width categories.}

\item{y}{A categorical variable defining the (vertical) segment categories.}

\item{width}{A numeric value to be summed across both categories.}
}
\value{
A Marimekko constructed with ggplot2.
}
\description{
A smarter stacked bar chart.
}
\examples{
hec <- dplyr::filter(data.frame(HairEyeColor), Sex == "Male")
marimekko(hec, Eye, Hair, Freq)
# Note: Compare to productplot::prodplot(
#   hec, Freq ~ Hair + Eye, mosaic("v")
# ) + aes(fill=Hair)
marimekko(hec, Hair, Eye, Freq) + labs(title = "Hair and Eye Color")
}

